Static Libraries and Zero Warnings

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -74,7 +74,7 @@ namespace Hazel {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 		// Rendering
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
@@ -90,5 +90,6 @@ namespace Hazel {
 	{
 		static bool show = true;
 		ImGui::ShowDemoWindow(&show);
+		
 	}
 }
